pyproject.toml
[project]
name = "datasette-examples"
version = "0.1a0"
description = "Load example SQL scripts into Datasette on startup"
readme = "README.md"
authors = [{name = "Simon Willison"}]
license = {text = "Apache-2.0"}
classifiers=[
    "Framework :: Datasette",
    "License :: OSI Approved :: Apache Software License"
]
requires-python = ">=3.8"
dependencies = [
    "datasette"
]

[project.urls]
Homepage = "https://github.com/datasette/datasette-examples"
Changelog = "https://github.com/datasette/datasette-examples/releases"
Issues = "https://github.com/datasette/datasette-examples/issues"
CI = "https://github.com/datasette/datasette-examples/actions"

[project.entry-points.datasette]
examples = "datasette_examples"

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-httpx"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
